pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "bambu-printer"
version = "0.1.0"
description = "Python client for Bambu Lab cloud API + MQTT"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "paho-mqtt>=2.0",
]

[project.optional-dependencies]
dev = ["ruff"]

[tool.setuptools]
packages = ["bambu"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
